FEAT: refactor login logic to use in registration
